+++ NEWS.md
@@ -1,3 +1,9 @@
+# stevethemes 0.2.0
+
+- Add `are_fonts_installed()` for further user diagnostics.
+- Add support for Merriweather font, which is a close cousin to the Palatino family.
+- Check for installation of Merriweather upon startup.
+
 # stevethemes 0.1.0
 
 - Initial developmental release